+++ src/components/FormItem.jsx
@@ -1,9 +1,7 @@
-import Button from './Button';
 import stringToSentenceCase from '../utilities/stringToSentenceCase';
 
 export default function FormItem({ type = 'text', dataAttributes, ...rest }) {
   const { id, value, name, onChange, children } = rest;
-  //   console.log(rest);
   return (
     <li className="form-item">
       <label htmlFor={id}>{stringToSentenceCase(name)}:</label>